Starting back over with version 1 and will rebuild up from there
after a long hiatus, going to try and pick this one back up again. Restarting from an earlier version because some of the later versions had a lot of experiments that were me not knowing what I was doing and stabbing in the dark. This version is super duper bad, so don't use it.